fixed: this should be an int, not a double

+++ opm/parser/eclipse/EclipseState/Schedule/MSW/SpiralICD.hpp
@@ -57,7 +57,7 @@ namespace Opm {
         double criticalValue() const;
         double widthTransitionRegion() const;
         double maxViscosityRatio() const;
-        double methodFlowScaling() const;
+        int methodFlowScaling() const;
 
         void updateScalingFactor(const double segment_length, const double completion_length);
         double scalingFactor() const;
